rename stuff and add link in readme
