+++ src/UI/Buyer/src/app/components/checkout/checkout/checkout.component.ts
@@ -41,8 +41,8 @@ import { ModalState } from 'src/app/models/shared.types'
 import { ErrorDisplayData, MiddlewareError } from 'src/app/models/error.types'
 import { Router } from '@angular/router'
 import { TranslateService } from '@ngx-translate/core'
-import { MooTrackService } from 'src/app/services/moosend.service'
-import { SitecoreCDPService } from 'src/app/services/sitecore-cdp.service'
+import { SitecoreSendTrackingService } from 'src/app/services/sitecore-send/sitecore-send-tracking.service'
+import { SitecoreCDPTrackingService } from 'src/app/services/sitecore-cdp/sitecore-cdp-tracking.service'
 
 @Component({
   templateUrl: './checkout.component.html',
@@ -100,8 +100,8 @@ export class OCMCheckout implements OnInit {
     private toastrService: ToastrService,
     private router: Router,
     private translate: TranslateService,
-    private mootrack: MooTrackService,
-    private cdp: SitecoreCDPService,
+    private mootrack: SitecoreSendTrackingService,
+    private cdp: SitecoreCDPTrackingService,
   ) {}
 
   async ngOnInit(): Promise<void> {
