+++ src/UI/Buyer/src/assets/appConfigs/defaultbuyer-uat.json
@@ -13,6 +13,9 @@
   "orderCloudApiUrl": "https://stagingapi.ordercloud.io",
   "useMoosend": false,
   "moosendWebsiteID": "xxxxxxxxxxxxxxxxxxxxxxxx",
+  "useSitecoreCDP": false,
+  "sitecoreCDPTargetEndpoint": "https://api-us.boxever.com/v1.2",
+  "sitecoreCDPApiClient": "xxxxxxxxxxxxxxxxxxxxxx",
   "theme": {
     "logoSrc": "assets/generic-logo.png"
   },
